refactor: split F1TelemetryDashboard into Container + Shell (7.1)
DashboardContainer   — UI state (theme/split/presets/feedback), all side-effects,
                        all event handlers; calls useDashboard(); renders DashboardShell.
DashboardShell       — pure presentational component; no hooks/side-effects.
                        Receives data + handlers as props — fully testable in isolation.
F1TelemetryDashboard — reduced to 10-line entry point: <DashboardContainer />.

Result:
  - Logic file (DashboardContainer): ~250 lines
  - UI file (DashboardShell): ~260 lines
  - Entry point: 10 lines
  vs. 813-line monolith before.
