+++ src/components/F1TelemetryDashboard.tsx
@@ -1,863 +1,12 @@
-import { Suspense, lazy, useCallback, useEffect, useMemo, useState } from 'react';
-import { useDashboardFilters } from '../hooks/useDashboardFilters';
-import type { DashboardFilterSnapshot } from '../hooks/useDashboardFilters';
-import { useDashboardSelectionData } from '../hooks/useDashboardSelectionData';
-import { useDashboardViewModel } from '../hooks/useDashboardViewModel';
-import { COLORS } from '../constants/colors';
-import { DriverProvider } from '../contexts/DriverContext';
-import {
-  useDrivers,
-  useIntervals,
-  useLapLocation,
-  useLapTelemetry,
-  useLaps,
-  useMeetings,
-  usePits,
-  usePositions,
-  useRaceControl,
-  useSessions,
-  useSessionResult,
-  useStints,
-  useTeamRadio,
-  useWeather,
-} from '../hooks/useOpenF1';
-import { ErrorBoundary } from './ErrorBoundary';
-import { DashboardHeader } from './dashboard/DashboardHeader';
-import { DashboardSelectors } from './dashboard/DashboardSelectors';
-import { DashboardTabs } from './dashboard/DashboardTabs';
-import { DriverSelector } from './dashboard/DriverSelector';
-import { TelemetryTab } from './dashboard/TelemetryTab';
-import { TrackMapTab } from './dashboard/TrackMapTab';
-import { ChartSkeleton, Err } from './dashboard/shared';
-import type { Tab } from './dashboard/types';
-
-type ThemeMode = 'dark' | 'light';
-
-type SavedPreset = {
-  name: string;
-  snapshot: DashboardFilterSnapshot;
-  splitMode: boolean;
-  themeMode?: ThemeMode;
-  savedAt: string;
-};
-
-const PRESET_STORAGE_KEY = 'f1-telemetry-dashboard:presets';
-const THEME_STORAGE_KEY = 'f1-telemetry-dashboard:theme';
-const MAX_DRIVER_SLOTS = 4;
-const HASH_SCROLL_MAX_ATTEMPTS = 12;
-const HASH_SCROLL_INTERVAL_MS = 120;
-const FEEDBACK_CLEAR_MS = 2600;
-const DEFAULT_IFRAME_HEIGHT = 920;
-const PANEL_IFRAME_HEIGHT = 720;
-const YEAR_OPTIONS = Array.from(
-  { length: new Date().getFullYear() - 2022 },
-  (_, index) => 2023 + index,
-);
-const TAB_LABELS = {
-  telemetry: 'Telemetry',
-  tires: 'Tyres & Strategy',
-  energy: 'DRS, Gears & RPM',
-  trackmap: 'Track Map',
-  positions: 'Race Positions',
-  intervals: 'Intervals',
-  radio: 'Team Radio',
-  incidents: 'Race Control',
-  weather: 'Weather',
-  broadcast: 'Broadcast',
-} as const;
-
-const StrategyTab = lazy(() => import('./dashboard/StrategyTab').then((module) => ({ default: module.StrategyTab })));
-const EnergyTab = lazy(() => import('./dashboard/EnergyTab').then((module) => ({ default: module.EnergyTab })));
-const RadioTab = lazy(() => import('./dashboard/RadioTab').then((module) => ({ default: module.RadioTab })));
-const IncidentsTab = lazy(() => import('./dashboard/IncidentsTab').then((module) => ({ default: module.IncidentsTab })));
-const WeatherTab = lazy(() => import('./dashboard/WeatherTab').then((module) => ({ default: module.WeatherTab })));
-const PositionsTab = lazy(() => import('./dashboard/PositionsTab').then((module) => ({ default: module.PositionsTab })));
-const IntervalsTab = lazy(() => import('./dashboard/IntervalsTab').then((module) => ({ default: module.IntervalsTab })));
-const BroadcastTab = lazy(() => import('./dashboard/BroadcastTab').then((module) => ({ default: module.BroadcastTab })));
-
-function TabLoadingPlaceholder({ label }: { label: string }) {
-  return (
-    <div className="dashboard-panel rounded-[16px] p-6 text-sm text-[color:var(--text-muted)] sm:rounded-[18px] sm:p-8">
-      <ChartSkeleton label={label} className="h-32" />
-    </div>
-  );
-}
-
-function readInitialSplitMode() {
-  if (typeof window === 'undefined') return false;
-  return new URLSearchParams(window.location.search).get('layout') === 'split';
-}
-
-function readInitialEmbedMode() {
-  if (typeof window === 'undefined') return false;
-  return new URLSearchParams(window.location.search).get('embed') === '1';
-}
-
-function normalizeThemeMode(value: string | null | undefined): ThemeMode | null {
-  if (value === 'light' || value === 'dark') return value;
-  return null;
-}
-
-function readInitialThemeMode(): ThemeMode {
-  if (typeof window === 'undefined') return 'dark';
-
-  const params = new URLSearchParams(window.location.search);
-  const fromQuery = normalizeThemeMode(params.get('theme'));
-  if (fromQuery) return fromQuery;
-
-  try {
-    const fromStorage = normalizeThemeMode(window.localStorage.getItem(THEME_STORAGE_KEY));
-    if (fromStorage) return fromStorage;
-  } catch {
-    // Ignore storage failures and fall back to the default theme.
-  }
-
-  return 'dark';
-}
-
-function readSavedPresets() {
-  if (typeof window === 'undefined') return {} as Record<string, SavedPreset>;
-  try {
-    const raw = window.localStorage.getItem(PRESET_STORAGE_KEY);
-    if (!raw) return {};
-    const parsed = JSON.parse(raw) as Record<string, SavedPreset>;
-    return parsed ?? {};
-  } catch {
-    return {};
-  }
-}
-
 /**
- * Builds shareable dashboard URLs from the persisted filter schema.
- * Query params are: year, circuit, session, drivers, lap, tab, optional layout,
- * optional embed=1, optional theme=light, and an optional panel hash anchor.
+ * F1TelemetryDashboard — entry point component.
+ *
+ * All data coordination lives in DashboardContainer.
+ * All UI rendering lives in DashboardShell.
  */
-function buildDashboardUrl(
-  snapshot: DashboardFilterSnapshot,
-  splitMode: boolean,
-  embedMode = false,
-  themeMode: ThemeMode = 'dark',
-  anchorId?: string,
-) {
-  if (typeof window === 'undefined') return '';
 
-  const params = new URLSearchParams();
-  params.set('year', String(snapshot.year));
-  if (snapshot.circuit) params.set('circuit', snapshot.circuit);
-  if (snapshot.sessionKey != null) params.set('session', String(snapshot.sessionKey));
-  if (snapshot.driverNums.length > 0) params.set('drivers', snapshot.driverNums.join(','));
-  params.set('lap', String(snapshot.lapNum));
-  params.set('tab', snapshot.tab);
-  if (splitMode) params.set('layout', 'split');
-  if (embedMode) params.set('embed', '1');
-  if (themeMode === 'light') params.set('theme', 'light');
-
-  const query = params.toString();
-  const hash = anchorId ? `#${anchorId}` : window.location.hash;
-  return `${window.location.origin}${window.location.pathname}${query ? `?${query}` : ''}${hash}`;
-}
-
-function buildIframeSnippet(
-  snapshot: DashboardFilterSnapshot,
-  splitMode: boolean,
-  themeMode: ThemeMode,
-  anchorId?: string,
-  iframeHeight = DEFAULT_IFRAME_HEIGHT,
-) {
-  const src = buildDashboardUrl(snapshot, splitMode, true, themeMode, anchorId);
-  const background = themeMode === 'light' ? COLORS.fallback.iframeLight : COLORS.fallback.iframeDark;
-  return [
-    `<iframe`,
-    `  src="${src}"`,
-    `  title="f1stories.gr F1 Telemetry Dashboard"`,
-    `  width="100%"`,
-    `  height="${iframeHeight}"`,
-    `  loading="lazy"`,
-    `  style="border:0; width:100%; max-width:100%; background:${background};"`,
-    `></iframe>`,
-  ].join('\n');
-}
-
-function getClipboardErrorMessage(error: unknown, label: string) {
-  const isDomException = typeof DOMException !== 'undefined' && error instanceof DOMException;
-  if (isDomException && (error.name === 'NotAllowedError' || error.name === 'SecurityError')) {
-    return `${label} clipboard denied; copy manually`;
-  }
-  return `${label} could not be copied; copy manually`;
-}
-
-function isShareCancel(error: unknown) {
-  return typeof DOMException !== 'undefined' && error instanceof DOMException && error.name === 'AbortError';
-}
-
-function useDriverLapData(sessionKey: number | null, driverNumber: number | null) {
-  return useLaps(sessionKey, driverNumber);
-}
-
-type DriverTelemetryWindow = {
-  lapStart: string | null;
-  nextLapStart: string | null;
-};
-
-function useDriverWindowData({
-  sessionKey,
-  driverNumber,
-  telemetryWindow,
-  needsTelemetryData,
-  needsLocationData,
-}: {
-  sessionKey: number | null;
-  driverNumber: number | null;
-  telemetryWindow?: DriverTelemetryWindow;
-  needsTelemetryData: boolean;
-  needsLocationData: boolean;
-}) {
-  const lapStart = telemetryWindow?.lapStart || null;
-  const nextLapStart = telemetryWindow?.nextLapStart || null;
-
-  const telemetry = useLapTelemetry(
-    sessionKey,
-    driverNumber,
-    needsTelemetryData ? lapStart : null,
-    needsTelemetryData ? nextLapStart : null,
-  );
-  const location = useLapLocation(
-    needsLocationData ? sessionKey : null,
-    driverNumber,
-    needsLocationData ? lapStart : null,
-    needsLocationData ? nextLapStart : null,
-  );
-
-  return { telemetry, location };
-}
+import { DashboardContainer } from './DashboardContainer';
 
 export default function F1TelemetryDashboard() {
-  const filters = useDashboardFilters();
-  const setLapNum = filters.setLapNum;
-  const needsTelemetryData = filters.tab === 'telemetry' || filters.tab === 'energy';
-  const needsStrategyData = filters.tab === 'tires';
-  const needsRadioData = filters.tab === 'radio';
-  const needsIncidentsData = filters.tab === 'incidents';
-  const needsWeatherData = filters.tab === 'weather';
-  const needsLocationData = filters.tab === 'trackmap';
-  const needsPositionsData = filters.tab === 'positions';
-  const needsIntervalsData = filters.tab === 'intervals';
-  const [splitMode, setSplitMode] = useState(readInitialSplitMode);
-  const [embedMode] = useState(readInitialEmbedMode);
-  const [themeMode, setThemeMode] = useState<ThemeMode>(readInitialThemeMode);
-  const [presetName, setPresetName] = useState('');
-  const [feedback, setFeedback] = useState<string | null>(null);
-  const [savedPresets, setSavedPresets] = useState<Record<string, SavedPreset>>(readSavedPresets);
-
-  const meetings = useMeetings(filters.year);
-  const sessions = useSessions(filters.year, filters.circuit);
-  const drivers = useDrivers(filters.sessionKey);
-  const sessionResults = useSessionResult(filters.sessionKey);
-  const driverSlots = useMemo(
-    () => Array.from({ length: MAX_DRIVER_SLOTS }, (_, index) => filters.driverNums[index] ?? null),
-    [filters.driverNums],
-  );
-
-  const lapStates = [
-    useDriverLapData(filters.sessionKey, driverSlots[0]),
-    useDriverLapData(filters.sessionKey, driverSlots[1]),
-    useDriverLapData(filters.sessionKey, driverSlots[2]),
-    useDriverLapData(filters.sessionKey, driverSlots[3]),
-  ];
-
-  const stints = useStints(needsStrategyData ? filters.sessionKey : null);
-  const pits = usePits(needsStrategyData ? filters.sessionKey : null);
-  const weather = useWeather(needsWeatherData ? filters.sessionKey : null);
-  const raceControl = useRaceControl(needsIncidentsData ? filters.sessionKey : null);
-  const teamRadio = useTeamRadio(needsRadioData ? filters.sessionKey : null);
-  const positions = usePositions(needsPositionsData ? filters.sessionKey : null);
-  const intervals = useIntervals(needsIntervalsData ? filters.sessionKey : null);
-
-  const selectionData = useDashboardSelectionData({
-    meetings: meetings.data,
-    sessions: sessions.data,
-    drivers: drivers.data,
-    sessionResults: sessionResults.data,
-    sessionResultsLoading: sessionResults.loading,
-    lapStates,
-    circuit: filters.circuit,
-    sessionKey: filters.sessionKey,
-    driverNums: filters.driverNums,
-    lapNum: filters.lapNum,
-    driverSelectionAuto: filters.driverSelectionAuto,
-    lapSelectionAuto: filters.lapSelectionAuto,
-    setCircuit: filters.setCircuit,
-    setSessionKey: filters.setSessionKey,
-    setDriverNums: filters.setAutoDriverNums,
-    setLapNum: filters.setAutoLapNum,
-  });
-
-  const driver0Data = useDriverWindowData({
-    sessionKey: filters.sessionKey,
-    driverNumber: driverSlots[0],
-    telemetryWindow: selectionData.telemetryWindows[0],
-    needsTelemetryData,
-    needsLocationData,
-  });
-  const driver1Data = useDriverWindowData({
-    sessionKey: filters.sessionKey,
-    driverNumber: driverSlots[1],
-    telemetryWindow: selectionData.telemetryWindows[1],
-    needsTelemetryData,
-    needsLocationData,
-  });
-  const driver2Data = useDriverWindowData({
-    sessionKey: filters.sessionKey,
-    driverNumber: driverSlots[2],
-    telemetryWindow: selectionData.telemetryWindows[2],
-    needsTelemetryData,
-    needsLocationData,
-  });
-  const driver3Data = useDriverWindowData({
-    sessionKey: filters.sessionKey,
-    driverNumber: driverSlots[3],
-    telemetryWindow: selectionData.telemetryWindows[3],
-    needsTelemetryData,
-    needsLocationData,
-  });
-
-  const telemetryStates = [
-    driver0Data.telemetry,
-    driver1Data.telemetry,
-    driver2Data.telemetry,
-    driver3Data.telemetry,
-  ];
-
-  const locationStates = [
-    driver0Data.location,
-    driver1Data.location,
-    driver2Data.location,
-    driver3Data.location,
-  ];
-  const locationByDriver = Object.fromEntries(
-    filters.driverNums.map((driverNum, index) => [driverNum, locationStates[index]?.data || null]),
-  ) as Record<number, ReturnType<typeof useLapLocation>['data']>;
-  const locationLoading = needsLocationData && locationStates.some((s) => s.loading);
-  const telemetryByDriver = Object.fromEntries(
-    filters.driverNums.map((driverNumber, index) => [driverNumber, telemetryStates[index]?.data || null]),
-  ) as Record<number, ReturnType<typeof useLapTelemetry>['data']>;
-
-  const viewModel = useDashboardViewModel({
-    activeTab: filters.tab,
-    allLaps: selectionData.allLaps,
-    driverMap: selectionData.driverMap,
-    driverNums: filters.driverNums,
-    lapNum: filters.lapNum,
-    lapOptions: selectionData.lapOptions,
-    stints: stints.data,
-    pits: pits.data,
-    weather: weather.data,
-    raceControl: raceControl.data,
-    teamRadio: teamRadio.data,
-    telemetryByDriver,
-  });
-  const driverContextValue = useMemo(
-    () => ({
-      driverNums: filters.driverNums,
-      driverMap: selectionData.driverMap,
-      driverColor: viewModel.driverColor,
-    }),
-    [filters.driverNums, selectionData.driverMap, viewModel.driverColor],
-  );
-
-  const anyLoading = meetings.loading || sessions.loading || drivers.loading || sessionResults.loading;
-  const lapsLoading = lapStates.some((state) => state.loading);
-  const telemetryLoading = needsTelemetryData && telemetryStates.some((state) => state.loading);
-  const totalLaps = selectionData.lapOptions.length > 0 ? selectionData.lapOptions[selectionData.lapOptions.length - 1] : null;
-  const currentLapIndex = selectionData.lapOptions.indexOf(filters.lapNum);
-  const canStepBackward = currentLapIndex > 0;
-  const canStepForward = currentLapIndex >= 0 && currentLapIndex < selectionData.lapOptions.length - 1;
-  const stepLap = useCallback((direction: -1 | 1) => {
-    const nextIndex = currentLapIndex + direction;
-    const nextLap = selectionData.lapOptions[nextIndex];
-    if (nextLap != null) {
-      setLapNum(nextLap);
-    }
-  }, [currentLapIndex, selectionData.lapOptions, setLapNum]);
-  const summaryPills = useMemo(() => {
-    return viewModel.lapSummaries.slice(0, 2).map((summary, index) => ({
-      label: `P${index + 1}`,
-      driver: summary.name,
-      detail: summary.gapToLeader != null && summary.gapToLeader > 0 ? `+${summary.gapToLeader.toFixed(3)}s` : summary.lapTime != null ? `${summary.lapTime.toFixed(3)}s` : '—',
-      tone: index === 0 ? 'blue' as const : 'purple' as const,
-    }));
-  }, [viewModel.lapSummaries]);
-  const quickChips = useMemo(() => {
-    const leader = viewModel.lapSummaries[0];
-    const pitLabel = needsStrategyData
-      ? pits.loading
-        ? 'Pits …'
-        : `Pits ${viewModel.filteredPits.length}`
-      : 'Pits —';
-    return [
-      { label: leader?.lapTime ? `Fastest ${leader.name} ${leader.lapTime.toFixed(3)}s` : `Focus L${filters.lapNum}`, tone: 'purple' as const },
-      { label: pitLabel, tone: 'amber' as const },
-      { label: `Drivers ${filters.driverNums.length}`, tone: 'blue' as const },
-      { label: `Laps ${selectionData.lapOptions.length}`, tone: 'neutral' as const },
-    ];
-  }, [filters.driverNums.length, filters.lapNum, needsStrategyData, pits.loading, selectionData.lapOptions.length, viewModel.filteredPits.length, viewModel.lapSummaries]);
-  const defaultPresetName = useMemo(() => {
-    const gp = filters.circuit || 'session';
-    return `${gp.toLowerCase().replace(/\s+/g, '-')}-lap-${filters.lapNum}`;
-  }, [filters.circuit, filters.lapNum]);
-  const presetNames = useMemo(
-    () => Object.values(savedPresets).sort((left, right) => right.savedAt.localeCompare(left.savedAt)).map((preset) => preset.name),
-    [savedPresets],
-  );
-  const sessionLabel = useMemo(
-    () => selectionData.sessionOptions.find((option) => option.v === filters.sessionKey)?.l ?? (filters.sessionKey != null ? `Session ${filters.sessionKey}` : 'Session'),
-    [filters.sessionKey, selectionData.sessionOptions],
-  );
-  const embedTitle = useMemo(() => {
-    const parts = [filters.circuit, sessionLabel].filter(Boolean);
-    return parts.length > 0 ? parts.join(' · ') : 'F1 Telemetry Embed';
-  }, [filters.circuit, sessionLabel]);
-  const embedSubtitle = useMemo(
-    () => `${filters.year} season · ${TAB_LABELS[filters.tab]} view`,
-    [filters.tab, filters.year],
-  );
-  const embedContext = useMemo(
-    () => [
-      { label: 'Lap', value: `L${filters.lapNum}` },
-      { label: 'Drivers', value: `${filters.driverNums.length}/${MAX_DRIVER_SLOTS}` },
-      { label: 'View', value: TAB_LABELS[filters.tab] },
-    ],
-    [filters.driverNums.length, filters.lapNum, filters.tab],
-  );
-  const openDashboardUrl = useMemo(
-    () => buildDashboardUrl(filters.snapshot, splitMode, false, themeMode),
-    [filters.snapshot, splitMode, themeMode],
-  );
-  const contentLayoutClass = splitMode
-    ? `grid ${embedMode ? 'gap-5 pb-8' : 'gap-4 pb-12 sm:gap-5 sm:pb-16'} xl:grid-cols-2 xl:[&>*:first-child]:col-span-2`
-    : `${embedMode ? 'space-y-5 pb-8' : 'space-y-4 pb-12 sm:space-y-5 sm:pb-16'}`;
-  const pageShellClass = embedMode
-    ? 'relative mx-auto max-w-[1320px] px-3 py-3 sm:px-4 sm:py-4'
-    : 'relative mx-auto max-w-[1440px] px-4 py-3 sm:px-6 sm:py-4';
-  const tabBoundaryResetKey = `${filters.tab}:${filters.sessionKey ?? 'none'}:${filters.lapNum}:${filters.driverNums.join(',')}`;
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return;
-
-    const url = buildDashboardUrl(filters.snapshot, splitMode, embedMode, themeMode);
-    window.history.replaceState({}, '', url);
-  }, [embedMode, filters.snapshot, splitMode, themeMode]);
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return;
-
-    const targetId = window.location.hash.replace(/^#/, '');
-    if (!targetId) return;
-
-    let attempts = 0;
-    const timerId = window.setInterval(() => {
-      const target = window.document.getElementById(targetId);
-      if (target) {
-        target.scrollIntoView({ behavior: 'smooth', block: 'start' });
-        window.clearInterval(timerId);
-        return;
-      }
-
-      attempts += 1;
-      if (attempts >= HASH_SCROLL_MAX_ATTEMPTS) {
-        window.clearInterval(timerId);
-      }
-    }, HASH_SCROLL_INTERVAL_MS);
-
-    return () => window.clearInterval(timerId);
-  }, [filters.tab]);
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return;
-    try {
-      window.localStorage.setItem(PRESET_STORAGE_KEY, JSON.stringify(savedPresets));
-    } catch {
-      // Ignore storage failures and keep the app usable.
-    }
-  }, [savedPresets]);
-
-  useEffect(() => {
-    if (!feedback || typeof window === 'undefined') return;
-
-    const timeoutId = window.setTimeout(() => setFeedback(null), FEEDBACK_CLEAR_MS);
-    return () => window.clearTimeout(timeoutId);
-  }, [feedback]);
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return;
-
-    const root = window.document.documentElement;
-    root.classList.toggle('theme-light', themeMode === 'light');
-    root.classList.toggle('theme-dark', themeMode === 'dark');
-
-    const themeColor = themeMode === 'light' ? COLORS.fallback.iframeLight : COLORS.fallback.iframeDark;
-    window.document.querySelector('meta[name="theme-color"]')?.setAttribute('content', themeColor);
-
-    try {
-      window.localStorage.setItem(THEME_STORAGE_KEY, themeMode);
-    } catch {
-      // Ignore storage failures and keep the app usable.
-    }
-  }, [themeMode]);
-
-  const handlePresetNameChange = useCallback((value: string) => {
-    setPresetName(value);
-    const preset = savedPresets[value];
-    if (!preset) return;
-
-    filters.applySnapshot(preset.snapshot);
-    setSplitMode(preset.splitMode);
-    if (preset.themeMode) {
-      setThemeMode(preset.themeMode);
-    }
-    setFeedback(`Loaded preset ${preset.name}`);
-  }, [filters, savedPresets]);
-
-  const handleSavePreset = useCallback(() => {
-    const name = presetName.trim() || defaultPresetName;
-    const preset: SavedPreset = {
-      name,
-      snapshot: filters.snapshot,
-      splitMode,
-      themeMode,
-      savedAt: new Date().toISOString(),
-    };
-
-    setSavedPresets((prev) => ({ ...prev, [name]: preset }));
-    setPresetName(name);
-    setFeedback(`Saved preset ${name}`);
-  }, [defaultPresetName, filters.snapshot, presetName, splitMode, themeMode]);
-
-  const shareSnapshot = useCallback(async (snapshot: DashboardFilterSnapshot, label: string) => {
-    const url = buildDashboardUrl(snapshot, splitMode, embedMode, themeMode);
-    let clipboardErrorMessage: string | null = null;
-    try {
-      if (navigator.clipboard?.writeText) {
-        await navigator.clipboard.writeText(url);
-        setFeedback(`${label} copied`);
-        return;
-      }
-    } catch (error) {
-      clipboardErrorMessage = getClipboardErrorMessage(error, label);
-    }
-
-    try {
-      if (navigator.share) {
-        await navigator.share({ title: `f1stories.gr ${TAB_LABELS[snapshot.tab]} view`, url });
-        setFeedback(`${label} shared`);
-        return;
-      }
-    } catch (error) {
-      if (!isShareCancel(error)) {
-        clipboardErrorMessage = clipboardErrorMessage ?? `${label} could not be shared; copy manually`;
-      }
-    }
-
-    window.prompt('Copy this link', url);
-    setFeedback(clipboardErrorMessage ?? `${label} ready`);
-  }, [embedMode, splitMode, themeMode]);
-
-  const embedSnapshot = useCallback(async (snapshot: DashboardFilterSnapshot, label: string) => {
-    const snippet = buildIframeSnippet(snapshot, splitMode, themeMode);
-    let clipboardErrorMessage: string | null = null;
-
-    try {
-      if (navigator.clipboard?.writeText) {
-        await navigator.clipboard.writeText(snippet);
-        setFeedback(`${label} copied`);
-        return;
-      }
-    } catch (error) {
-      clipboardErrorMessage = getClipboardErrorMessage(error, label);
-    }
-
-    window.prompt('Copy this iframe snippet', snippet);
-    setFeedback(clipboardErrorMessage ?? `${label} ready`);
-  }, [splitMode, themeMode]);
-
-  const handleShare = useCallback(async () => {
-    await shareSnapshot(filters.snapshot, 'Share link');
-  }, [filters.snapshot, shareSnapshot]);
-
-  const handleShareTab = useCallback(async (tab: Tab) => {
-    await shareSnapshot({ ...filters.snapshot, tab }, `${TAB_LABELS[tab]} link`);
-  }, [filters.snapshot, shareSnapshot]);
-
-  const handleEmbed = useCallback(async () => {
-    await embedSnapshot(filters.snapshot, 'Embed code');
-  }, [embedSnapshot, filters.snapshot]);
-
-  const handleEmbedTab = useCallback(async (tab: Tab) => {
-    await embedSnapshot({ ...filters.snapshot, tab }, `${TAB_LABELS[tab]} embed`);
-  }, [embedSnapshot, filters.snapshot]);
-
-  const handleEmbedPanel = useCallback(async (panelId: string) => {
-    const snippet = buildIframeSnippet(filters.snapshot, false, themeMode, panelId, PANEL_IFRAME_HEIGHT);
-    let clipboardErrorMessage: string | null = null;
-
-    try {
-      if (navigator.clipboard?.writeText) {
-        await navigator.clipboard.writeText(snippet);
-        setFeedback('Panel embed copied');
-        return;
-      }
-    } catch (error) {
-      clipboardErrorMessage = getClipboardErrorMessage(error, 'Panel embed');
-    }
-
-    window.prompt('Copy this iframe snippet', snippet);
-    setFeedback(clipboardErrorMessage ?? 'Panel embed ready');
-  }, [filters.snapshot, themeMode]);
-
-  const handlePrint = useCallback(() => {
-    setFeedback('Opening print dialog');
-    window.print();
-  }, []);
-
-  const handleToggleSplit = useCallback(() => {
-    setSplitMode((prev) => !prev);
-    setFeedback(splitMode ? 'Split layout disabled' : 'Split layout enabled');
-  }, [splitMode]);
-
-  const handleToggleTheme = useCallback(() => {
-    const nextTheme = themeMode === 'light' ? 'dark' : 'light';
-    setThemeMode(nextTheme);
-    setFeedback(nextTheme === 'light' ? 'Light mode enabled' : 'Dark mode enabled');
-  }, [themeMode]);
-
-  const handleBack = useCallback(() => {
-    if (window.history.length > 1) {
-      window.history.back();
-    } else {
-      setFeedback('No previous page in history');
-    }
-  }, []);
-
-  return (
-    <div className={`dashboard-app ${embedMode ? '' : 'min-h-screen'} ${themeMode === 'light' ? 'theme-light' : 'theme-dark'} ${embedMode ? 'embed-mode' : ''}`}>
-      <div className={pageShellClass}>
-        {!embedMode && (
-          <a
-            href="#main-content"
-            className="sr-only focus:not-sr-only focus:fixed focus:left-2 focus:top-2 z-50 rounded bg-white px-4 py-2 text-sm font-semibold text-black shadow-lg"
-          >
-            Skip to content
-          </a>
-        )}
-
-        <DashboardHeader
-          loading={anyLoading}
-          presetName={presetName}
-          presetNames={presetNames}
-          feedback={feedback}
-          splitMode={splitMode}
-          embedMode={embedMode}
-          themeMode={themeMode}
-          embedTitle={embedTitle}
-          embedSubtitle={embedSubtitle}
-          embedContext={embedContext}
-          openDashboardUrl={openDashboardUrl}
-          onPresetNameChange={handlePresetNameChange}
-          onSavePreset={handleSavePreset}
-          onShare={handleShare}
-          onEmbed={handleEmbed}
-          onPrint={handlePrint}
-          onToggleSplit={handleToggleSplit}
-          onToggleTheme={handleToggleTheme}
-          onBack={handleBack}
-        />
-
-        <main id="main-content" tabIndex={-1}>
-          <DashboardSelectors
-            year={filters.year}
-            circuit={filters.circuit}
-            sessionKey={filters.sessionKey}
-            lapNum={filters.lapNum}
-            totalLaps={totalLaps}
-            yearOptions={YEAR_OPTIONS}
-            circuitOptions={selectionData.circuitOptions}
-            sessionOptions={selectionData.sessionOptions}
-            lapOptions={selectionData.lapOptions}
-            meetingsLoading={meetings.loading}
-            sessionsLoading={sessions.loading}
-            lapsLoading={lapsLoading}
-            summaryPills={summaryPills}
-            quickChips={quickChips}
-            canStepBackward={canStepBackward}
-            canStepForward={canStepForward}
-            embedMode={embedMode}
-            onYearChange={filters.handleYearChange}
-            onCircuitChange={filters.handleCircuitChange}
-            onSessionChange={filters.handleSessionChange}
-            onLapChange={setLapNum}
-            onStepLap={stepLap}
-          />
-
-          {meetings.error && <Err msg={`Failed to load calendar: ${meetings.error}`} onAction={meetings.refetch} />}
-          {sessions.error && <Err msg={`Failed to load sessions: ${sessions.error}`} onAction={sessions.refetch} />}
-          {drivers.error && <Err msg={`Failed to load drivers: ${drivers.error}`} onAction={drivers.refetch} />}
-
-          {!embedMode && (
-            <DriverSelector
-              drivers={selectionData.driverList}
-              selectedDrivers={filters.driverNums}
-              embedMode={embedMode}
-              onToggle={filters.toggleDriver}
-            />
-          )}
-
-          <DashboardTabs
-            activeTab={filters.tab}
-            onChange={filters.setTab}
-            embedMode={embedMode}
-            onShareTab={handleShareTab}
-            onEmbedTab={handleEmbedTab}
-          />
-
-          <DriverProvider {...driverContextValue}>
-            <ErrorBoundary label={TAB_LABELS[filters.tab]} resetKey={tabBoundaryResetKey}>
-              <div className={contentLayoutClass}>
-                {filters.tab === 'telemetry' && (
-                  <TelemetryTab
-                    lapNum={filters.lapNum}
-                    lapsLoading={lapsLoading}
-                    sectorRows={viewModel.sectorRows}
-                    telemetryLoading={telemetryLoading}
-                    telemetryError={telemetryStates[0].error}
-                    telemetryPoints={telemetryStates[0].data?.length || 0}
-                    speedData={viewModel.speedData}
-                    comparisonSpeedData={viewModel.comparisonSpeedData}
-                    comparisonControlData={viewModel.comparisonControlData}
-                    lapTimeData={viewModel.lapTimeData}
-                    lapDeltaData={viewModel.lapDeltaData}
-                    lapSummaries={viewModel.lapSummaries}
-                    embedMode={embedMode}
-                    onEmbedPanel={handleEmbedPanel}
-                    onTelemetryRetry={telemetryStates[0].refetch}
-                  />
-                )}
-
-                {filters.tab === 'tires' && (
-                  <Suspense fallback={<TabLoadingPlaceholder label="Loading strategy view..." />}>
-                    <StrategyTab
-                      lapNum={filters.lapNum}
-                      stintsLoading={stints.loading}
-                      stintsByDriver={viewModel.stintsByDriver}
-                      pitsLoading={pits.loading}
-                      filteredPits={viewModel.filteredPits}
-                      embedMode={embedMode}
-                      onEmbedPanel={handleEmbedPanel}
-                    />
-                  </Suspense>
-                )}
-
-                {filters.tab === 'energy' && (
-                  <Suspense fallback={<TabLoadingPlaceholder label="Loading energy view..." />}>
-                    <EnergyTab
-                      lapNum={filters.lapNum}
-                      speedData={viewModel.speedData}
-                      comparisonEnergyData={viewModel.comparisonEnergyData}
-                      lapSummaries={viewModel.lapSummaries}
-                      telemetryLoading={telemetryLoading}
-                      embedMode={embedMode}
-                      onEmbedPanel={handleEmbedPanel}
-                    />
-                  </Suspense>
-                )}
-
-                {filters.tab === 'radio' && (
-                  <Suspense fallback={<TabLoadingPlaceholder label="Loading radio view..." />}>
-                    <RadioTab
-                      loading={teamRadio.loading}
-                      error={teamRadio.error}
-                      messages={viewModel.filteredRadio}
-                      onRetry={teamRadio.refetch}
-                    />
-                  </Suspense>
-                )}
-
-                {filters.tab === 'incidents' && (
-                  <Suspense fallback={<TabLoadingPlaceholder label="Loading race control view..." />}>
-                    <IncidentsTab
-                      loading={raceControl.loading}
-                      error={raceControl.error}
-                      messages={viewModel.raceControlMessages}
-                      onRetry={raceControl.refetch}
-                    />
-                  </Suspense>
-                )}
-
-                {filters.tab === 'weather' && (
-                  <Suspense fallback={<TabLoadingPlaceholder label="Loading weather view..." />}>
-                    <WeatherTab
-                      loading={weather.loading}
-                      error={weather.error}
-                      latestWeather={viewModel.latestWeather}
-                      sampleCount={weather.data?.length || 0}
-                      weatherRadar={viewModel.weatherRadar}
-                      weatherTrend={viewModel.weatherTrend}
-                      embedMode={embedMode}
-                      onEmbedPanel={handleEmbedPanel}
-                      onRetry={weather.refetch}
-                    />
-                  </Suspense>
-                )}
-
-                {filters.tab === 'trackmap' && (
-                  <TrackMapTab
-                    lapNum={filters.lapNum}
-                    locationByDriver={locationByDriver}
-                    locationLoading={locationLoading}
-                    embedMode={embedMode}
-                    onEmbedPanel={handleEmbedPanel}
-                  />
-                )}
-
-                {filters.tab === 'positions' && (
-                  <Suspense fallback={<TabLoadingPlaceholder label="Loading race positions..." />}>
-                    <PositionsTab
-                      positions={positions.data}
-                      positionsLoading={positions.loading}
-                      lapNum={filters.lapNum}
-                      embedMode={embedMode}
-                      onEmbedPanel={handleEmbedPanel}
-                    />
-                  </Suspense>
-                )}
-
-                {filters.tab === 'intervals' && (
-                  <Suspense fallback={<TabLoadingPlaceholder label="Loading interval data..." />}>
-                    <IntervalsTab
-                      intervals={intervals.data}
-                      intervalsLoading={intervals.loading}
-                      embedMode={embedMode}
-                      onEmbedPanel={handleEmbedPanel}
-                    />
-                  </Suspense>
-                )}
-
-                {filters.tab === 'broadcast' && (
-                  <Suspense fallback={<TabLoadingPlaceholder label="Loading broadcast view..." />}>
-                    <BroadcastTab
-                      lapNum={filters.lapNum}
-                      lapsLoading={lapsLoading}
-                      sectorRows={viewModel.sectorRows}
-                      lapSummaries={viewModel.lapSummaries}
-                      embedMode={embedMode}
-                      onEmbedPanel={handleEmbedPanel}
-                    />
-                  </Suspense>
-                )}
-              </div>
-            </ErrorBoundary>
-          </DriverProvider>
-        </main>
-      </div>
-    </div>
-  );
+  return <DashboardContainer />;
 }
